requirements.txt
opencv-python
numpy
mediapipe
tensorflow
pygame
gtts
mutagen
pytesseract
Pillow